jaeger: add install unit tests
Manifests are very prone to go wrong as Helm indentation
changes, etc are very easy to go wrong. This PR adds
unit tests to the render logic, so that changes to the
output manifests are tracked.

This follows the same pattern as that of other render
unit tests in core cp, viz, etc.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>