Fix linkerd check --proxy failure on Jobs (#6155)
When `linkerd check --proxy` is run in a given namespace, the check
command ensures that all pods within that namespace have a running proxy
and that and the pod is in a `Running` status.

This however, is not necessarily true when dealing with CronJobs or Jobs
as these resources create pods that run their containers for a short period
of time and then terminate. When these pods terminate, they get placed
in a `Completed` phase and their sidecars are no longer running. This
causes an issue in `linkerd check --proxy` because that command expects
that pods will always be in a running state.

This change effectively skips checking the proxy status of a pod if that
pod is in a `Completed` phase as a pod no longer has an actively running
proxy

Fixes #6128

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>