Create gateway-port option for mc link (#6175)
In case of a NodePort service type, there might be non-kubernetes load
balancers that map a public port to the selected NodePorts. This is not
something that is detectable, so a flag needs to be provided to specify
the public port during the `mc link` command.

Note that the probe port can already be set using `--set
gateway.probe.port 1234`.

Signed-off-by: Peter Smit <peter.smit@inscripta.io>