web: Update multicast-dns for CVE-2021-23386
Our web development environment has a dependency on webpack-dev-server,
which pulls in bonjour, which pulls in multicast-dns, which depends on
dns-packet, which has a [critical security vulnerability][1].

webpack/webpack-dev-server#3340 tracks fixing this dependency, but it
appears that the bonjour project is no longer maintained.

This change works around this issue by patching the multicast-dns
dependency to pull in a fixed version. This could potentially break mdns
functionality in our development environment, but we probably don't even
use this functionality.

Dependencies are bad.

[1]: https://nvd.nist.gov/vuln/detail/CVE-2021-23386