artifacthub-repo.yml file to claim charts ownership
The owners in that file match the entries in MAINTAINERS.md.
On the next release build this will upload that file into our Helm repo,
and we'll be able to claim ownership of the chart at artifacthub.io.

Once we do that we'll get a `repositoryID` we can add into that file, to
obtain the "Verified publisher" badge.

This change also removes a couple of no longer relevant OWNERS files.