The CP readiness checks were ignoring Destination (#6216)
Supersedes #6196

Waaaay back when the `destination` container was moved out of the
"public-api" pod, we failed to add the new pod into the list of pods to
be checked in the `control plane pods are ready` health check.

`linkerd viz install` uses that check before proceeding to install. So
if that happens too quickly we risk trying to run viz components without
a `destination` pod being ready, which was causing the viz pods to
fail and restart sporadically.

Most integration tests use `test/integration/install_test.go` to setup
the cluster, which calls `kubectl rollout status` after the core
install, which blocks on full readiness before installing viz.

But the `uninstall` test does its own install and doesn't have that
rollout blocking, so that's why we were seeing those pod restarts only
for that test!