Add support for legacy names in extension uninstall (#6226)
Because extension label values have been changed, an older install of an extension will fail to be uninstalled with a more recent CLI version. This is because the uninstall command checks the new label value which doesn't match older installs.

This small change fixes that by checking both -- the new, non-prefixed version and the old prefixed one. The bug should have a relatively small impact since there are only a few edge releases that contain the prefixed label value (an exception being stable-2.10.0 which used the prefixed version still).

Signed-off-by: Matei David <matei@buoyant.io>
Signed-off-by: jasonmorgan <jmorgan@f9vs.com>