Fix namespace always showing up in topology graph (#6236)
* Fix namespace always showing up in topology graph

Fixes #6211

In #6091, code was added to include the the namespace in the list of all
stat resource types. This was added to so that we'd have a complete list
of resources types that could be suggested by the CLI autocompletion
code. However, this list was also used by the web frontend in a query
that gathered metrics from all resource types. This then caused the
query to inadvertently create an inbound metric for deployment that
showed traffic from given namespace.

This change, breaks up the code so that we have a separate list for the
autocompletion code without the namespace value and the original list
used by the web frontend prior to #6091.

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>