proxy: v2.147.0 (#6243)
This release fixes a problem with the HTTP body buffering that was added
to support gRPC retries. The proxy would buffer all request bodies,
regardless of size or retry configurations. This has been fixed so that
only requests with a retry configuration are buffered (and only when
their bodies are less than 64KB).

This release also fixes an issue with the outbound ingress-mode proxy
where forwarded HTTP traffic could fail to detect when the target pod
was deleted, retrying connections forever. This only impacted traffic
forwarded directly to pod IPs (and not load balanced services). This has
been fixed temporarily by adding a failfast layer that triggers 502
errors when the endpoint has disconected, which cause the connection to
be torn down so that the ingress may reconnect. A more robust solution
will replace this in the future.

Furthermore, core dependencies have been updated including: futures,
hyper, socket2, and tokio.

---

* Fix MacOS conditional build in telemetry::process (linkerd/linkerd2-proxy#1023)
* deps: update `futures` to 0.3.15 (linkerd/linkerd2-proxy#1022)
* tracing: Split HTML-formatting into admin module (linkerd/linkerd2-proxy#1025)
* tracing: Simplify initialization (linkerd/linkerd2-proxy#1026)
* Repace linkerd-drain with drain from crates.io (linkerd/linkerd2-proxy#1027)
* app: Move the admin server into a subcrate (linkerd/linkerd2-proxy#1028)
* inbound: Simplify protocol-detection skipping (linkerd/linkerd2-proxy#1031)
* proxy-api: Update proxy-api to use the main branch (linkerd/linkerd2-proxy#1029)
* outbound: don't double-wrap replay bodies (linkerd/linkerd2-proxy#1036)
* ingress: Add failfast to the forwarder (linkerd/linkerd2-proxy#1035)
* Update tokio, hyper, and socket2 (linkerd/linkerd2-proxy#1037)
* Implement reconnect as a NewService (linkerd/linkerd2-proxy#1032)
* Introduce the tonic-watch crate (linkerd/linkerd2-proxy#1034)
* service-profiles: Wrap receiver types (linkerd/linkerd2-proxy#1038)
* retry: only wrap bodies when a request can be retried (linkerd/linkerd2-proxy#1039)