Allow adding to `values.yaml` without having to add to the `Values` struct
Addresses in part #6131

## The Problem

Sometimes people want to add to `values.yaml` stuff they can use solely in the Helm templates. Currently we have to remind them they need to update the `Values` structs as well.

## The Solution

This modifies a bit the logic behind the extraction of default field values and how we merge overrides to compute the final values in `linkerd install/upgrade`, so that entries in `values[-ha].yaml` that don't have a counter-part in the `Values` struct are also taken into account.

Note that if people want to make use of those values in the CLI code, they still need to add them to the `Values` struct for the config object to expose them.

## Implementation

- Modified `install.go`'s logic, moving some of `render()`'s logic into `getChart()`, whose sole purpose is to build Helm's `chart` object, setting `chart.Values` with the default values from `values.yaml`, and merging the default values of `values-ha.yaml` if required.
Then we merge into `chart.Values` the overrides from the flags (including the --set-* flags), similar to how we did it before.
- `renderOverrides()` now receives these default values, as it's no longer sufficient to retrieve them with `l5dcharts.NewValues()` (because that only considered entries from the `Values` struct)

## Other related changes

- Uncommented all the entries in `values.yaml` that were commented, so their proper default value is picked (even if that value is empty).
- Added `enabledPodAntiAffinity: false` into `values.yaml` for the same reason.
- Removed `grafanaUrl` and `prometheusUrl` from the `Values` struct, which are no longer used.
- Added a blurb in `BUILD.md` describing how additions to `values.yaml` can be made.

## Followups

- Similar change for the multicluster extension.
- Refactor the `TestRender()` unit test in `install_test.go` so that it also takes into account `values[-ha].yaml` entries not declared under the `Values` struct.