artifacthub.io Verified Publisher
Followup to #6181

After having uploaded the `artifacthub-repo.yml` file, we were able to
claim ownership of the linkerd2-edge repo. This generated a
`repositoryID` in artifacthub.io that we're adding into this same file,
so that we can be marked as "Verified Publisher".

Also we moved that file one level up into `charts/` so it doesn't get
included in the built chart. And created two versions of that file, one
for edge and another for stable (whose ID will be received when the next
  stable helm release is pushed), that will be copied into the helm repo
  by the `release.yml` workflow.