Add endpoint in profile response for pod DNS lookups
When we send a profile request with a pod IP, we get back an endpoint as
part of the response. This has two advantages: we avoid building a load
balancer and we can treat endpoint failure differently (with more of a
fail fast approach). At the moment, when we use a pod DNS as the target
of the profile lookup, we don't have an endpoint returned in the
response.

Through this change, the behaviour will be consistent. Whenever we look
up a pod (either through IP or DNS name) we will get an endpoint back.
The change also attempts to simplify some of the logic in GetProfile.

Signed-off-by: Matei David <matei@buoyant.io>