use `controller/k8s` pkg for caching instead
This commit updates the pre-fetching logic to instead use
`controller/k8s` which provides a wrapper around `pkg/k8s` with
caching by using shared informers underneath!

This means that we can use it, instead of manually fetching and
storing the kubernetes objects.

We had to use `k8s.NewAPI` for the initialization to take-in
`hc.kubeAPI` so as to support unit testing, etc as `hc.kubeAPI`
is how we pass the fake resources in unit tests.

`kubeAPI.Sync` fn also had to be updated as it fails to check
if the sp and ts sharedinformers are nil, which might be the
case in cases like this where they are not needed.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>