Add missing ignoreInboundports to CNI config
This change fixes an issue where tap does not work when running Linkerd
through Linkerd CNI installed via helm charts. This issue was caused by
the CNI chart's value not including tap control and admin ports in the
config. This caused tap request traffic to go to the inbound side of the
proxy as opposed to the respective tap control port.

Fixes #6224

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>