Add @mateiidavid to maintainers
@mateiidavid has a solid history of contributing to Linkerd over the
past six+ months. Notably, he led a project to introduce StatefulSet
support for the `linkerd-multicluster` extension comprising changes
across the stack from the service-mirror and destination controllers to
the proxy. Furthermore, he's been an active contributor to code reviews
and support discussions on GitHub and the Linkerd Slack.

This change adds @mateiidavid to `MAINTAINERS.md`. Once this change is
merged, he'll also be added to the relevant GitHub groups so that he has
approval and merge privileges on Linkerd repos.