Add integration test for servers bound to localhost (#6536)
TestLocalhostServer creates an nginx deployment which listens on localhost
and a slow-cooker which attempts to send traffic to the nginx.  Since
slow-cooker should not be able to connect to nginx's localhost address,
these requests should fail.

This test currently asserts that these requests should succeed.  Once https://github.com/linkerd/linkerd2/issues/6495 is fixed, this test should be updated to assert that these requests should fail.

Signed-off-by: Alex Leong <alex@buoyant.io>