check: Warn on SMI resources without `linkerd-smi` extension
Part of https://github.com/linkerd/linkerd2/issues/6647

This PR adds a new check which checks for the presence of SMI resources
and the absence of `linkerd-smi` extension and warns the user about
the deprecation in that case.

This also introduces a new `linkerd-deprecation` checker where we can
add any kind of checks to warn users about deprecations in the future.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>