docker: Parellize CLI builds
We currently build all of our CLI binaries serially, but if we have a
docker stage for each platform, we can parellize builds for each
platform, reducing build times significantly.

This change renames cli/Dockerfile-bin to cli/Dockerfile (so that we get
syntax highlighting in editors, etc) and restructures the Dockerfile to
have a docker stage for each platform. Then, there are two final stages:
'basic' and 'multi-arch'. The `bin/docker-build-cli-bin` utility
typically only builds the `basic` target; when `DOCKER_MULTIARCH` is
set, it also builds the target that includes arm binaries.