proxy: v2.154.0
This release improves policy handling for HTTP connections so that
requests are failed with a 403 Forbidden status (or a PERMISSION_DENIED
grpc-status, if appropriate).

Inbound metrics now include labels indicating the server and/or
authorization used to allow a connection or request to the proxy. Error
metrics now include an `unauthorized` error reason for traffic that is
denied by policy.

Finally, the outbound proxy no longer initializes mTLS or HTTP/2
upgrades when the target proxy is itself. This is done in preparation
for changes that will allow the proxy to stop forwarding connections on
`localhost` so that servers bound only on the loopback interface are not
exposed by Linkerd.

---

* build(deps): bump h2 from 0.3.3 to 0.3.4 (linkerd/linkerd2-proxy#1212)
* build(deps): bump libc from 0.2.99 to 0.2.100 (linkerd/linkerd2-proxy#1213)
* Use `ExtractParam` in transport metrics (linkerd/linkerd2-proxy#1211)
* policy: Add support for cluster-scoped default policies (linkerd/linkerd2-proxy#1210)
* Expose policy labels on inbound transport metrics (linkerd/linkerd2-proxy#1215)
* inbound: Expose permit labels on HTTP metrics (linkerd/linkerd2-proxy#1216)
* build(deps): bump tokio from 1.10.0 to 1.10.1 (linkerd/linkerd2-proxy#1218)
* build(deps): bump codecov/codecov-action from 2.0.2 to 2.0.3 (linkerd/linkerd2-proxy#1217)
* build(deps): bump hyper from 0.14.11 to 0.14.12 (linkerd/linkerd2-proxy#1221)
* build(deps): bump bytes from 1.0.1 to 1.1.0 (linkerd/linkerd2-proxy#1222)
* inbound: Return HTTP-level authorization errors (linkerd/linkerd2-proxy#1220)
* Skip TLS and H2 when target is inbound IP (linkerd/linkerd2-proxy#1219)